apply review suggestions

+++ app/jobs/rdv_solidarites_webhooks/process_agent_job.rb
@@ -14,17 +14,13 @@ module RdvSolidaritesWebhooks
     end
 
     def agent
-      @agent = Agent.find_by(rdv_solidarites_agent_id: @data[:id])
+      @agent ||= Agent.find_by(rdv_solidarites_agent_id: @data[:id])
     end
 
     def upsert_or_delete_agent
-      return delete_agent if event == "destroyed"
+      return agent.destroy! if event == "destroyed"
 
       UpsertRecordJob.perform_async("Agent", @data, { last_webhook_update_received_at: @meta[:timestamp] })
-    end
-
-    def delete_agent
-      agent.destroy!
     end
   end
 end
